+++ src/tooling/piral-cli/src/common/index.ts
@@ -27,3 +27,4 @@ export * from './scaffold';
 export * from './scripts';
 export * from './spec';
 export * from './template';
+export * from './url';